rename mirserverhooks to mir_server_hooks
