+++ src/miroil/CMakeLists.txt
@@ -13,7 +13,7 @@ add_library(miroil SHARED
     event_builder.cpp ${miroil_include}/miroil/event_builder.h
     input_device.cpp ${miroil_include}/miroil/input_device.h
     input_device_observer.cpp ${miroil_include}/miroil/input_device_observer.h
-    mirserverhooks.cpp ${miroil_include}/miroil/mirserverhooks.h
+    mir_server_hooks.cpp ${miroil_include}/miroil/mir_server_hooks.h
     prompt_session_listener.cpp ${miroil_include}/miroil/prompt_session_listener.h
     prompt_session_manager.cpp ${miroil_include}/miroil/prompt_session_manager.h    
     ${miroil_include}/miroil/display_configuration_storage.h
